+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## Current (in progress)
 
-- Initial version
+- Add endpoint to stream a CSV response [#5](https://github.com/etalab/api-tabular/pull/5)